Day 3
today we went through the regular expressions and everything they do. We used them to grab the data we put inside of the url and we then placed it into different inputs. I also had to go back and fix my code for some reason it broke it worked last time I was on it then I fixed it.